When in doubt, cast

+++ src/main/java/com/dsh105/nexus/command/module/bukkit/TenJavaCommand.java
@@ -129,7 +129,7 @@ public class TenJavaCommand extends CommandModule {
                             total += judge.getAssignedItems();
                             completed += judge.getCompletedItems();
                         }
-                        totalPercentCompleted = ((double) completed) / total;
+                        totalPercentCompleted = (int)(((double) completed) / total);
                         totalPercentCompleted *= 100;
                         event.respondWithPing("ten.java Judging stats: {0}/{1} items completed ({2})", completed + "", total + "", (totalPercentCompleted / judges.length) + "%");
                         return true;
